+++ packages/core/src/docs/examples/row-hover.stories.tsx
@@ -8,8 +8,8 @@ import {
     useAllMockedKinds,
 } from "../../data-editor/stories/utils.js";
 import type { GetRowThemeCallback } from "../../internal/data-grid/data-grid-render.js";
-import type { GridMouseEventArgs } from "../../internal/data-grid/data-grid-types.js";
 import { SimpleThemeWrapper } from "../../stories/story-utils.js";
+import type { GridMouseEventArgs } from "../../internal/data-grid/event-args.js";
 
 export default {
     title: "Glide-Data-Grid/DataEditor Demos",
